fix: check and fix some ABI marks before 7.24
Check if ABI flags for FUSE protocol payload headers, operation codes, constants and related structures are
matched with the protocol version, and fix unmatched ones.

Besides, this commit implements (and removes) some C macros marked with `TODO: re-define it` in
`protocol.rs`. Some are implemented as functions, and `FUSE_DEV_IOC_CLONE`
is implemented as a constant.

The versions this commit involved are between 7.9 and 7.23 (included).